refactor: improve app folder structure and namings
place reducers into separate folders for each reducer, so far there is
only one but I expect it will grow

rename the urls so they are upper-case as the naming convetions dictate
to keep everything readable and with logic
