+++ frontend/src/assets/static_data/StaticData.js
@@ -1,11 +1,11 @@
 export const STATIC_DATA = {
 	URL: {
-		loginPage: 'http://localhost:3009/login',
-		charts: 'http://localhost:3009/Products',
-		mainPage: 'http://localhost:3009/',
-		loginUser: 'http://localhost:3027/users/login',
-		fetchUser: 'http://localhost:3027/users/users',
-		fetchUsers: 'http://localhost:3027/users',
-		fetchData: 'http://localhost:3027/analytics',
+		LOGIN_PAGE: 'http://localhost:3009/login',
+		CHARTS: 'http://localhost:3009/Products',
+		MAINPAGE: 'http://localhost:3009/',
+		LOGIN_USER: 'http://localhost:3027/users/login',
+		FETCH_USER: 'http://localhost:3027/users/users',
+		FETCH_USERS: 'http://localhost:3027/users',
+		FETCH_DATA: 'http://localhost:3027/analytics',
 	}
 }
